test:unit script added, removed demo.test.js in test folder
