+++ jest.config.js
@@ -2,8 +2,7 @@ module.exports = {
     testPathIgnorePatterns: ["/.next/", "/node_modules/"],
     setupFilesAfterEnv: ["./setupTests.js"],
     verbose: true,
-    roots: ["<rootDir>/test"],
-    testMatch: ["**/test/**/*.e2e.js", "**/test/**/*.test.js"],
+    roots: ["<rootDir>/test", "<rootDir>/renderer"],
     setupFiles: ["<rootDir>/test/lib/setup.js"],
     testSequencer: "<rootDir>/test/lib/sequencer.js",
 };
